requirements.txt
streamlit
transformers
sentence-transformers
PyMuPDF
faiss-cpu
langchain
torch
